baseType configed and time_0->null
