+++ demo/src/test/java/io/xream/x7/AppTest.java
@@ -37,7 +37,7 @@ public class AppTest {
 //        xxxTest.domain();
 
 //        xxxTest.inOrder();
-        xxxTest.testOrderFind();
+//        xxxTest.testOrderFind();
 //        xxxTest.testOrderFindByAlia();
 
 //        xxxTest.testReyClient();
@@ -79,7 +79,7 @@ public class AppTest {
 
 //        xxxTest.testOne();
 //        xxxTest.testListCriteria();
-//        xxxTest.refreshByCondition();
+        xxxTest.refreshByCondition();
 //        xxxTest.testListCriteria();
 //        xxxTest.testRemove();
 //        xxxTest.testListCriteria();
