+++ x7-common/src/main/java/io/xream/x7/common/bean/SqlConditionCriteria.java
@@ -16,8 +16,9 @@
  */
 package io.xream.x7.common.bean;
 
+import io.xream.x7.common.util.BaseTypeUtil;
 import io.xream.x7.common.util.BeanUtil;
-import io.xream.x7.common.util.BeanUtilX;
+import io.xream.x7.common.util.NumberValueAndDateUtil;
 import io.xream.x7.common.util.StringUtil;
 
 import java.util.Iterator;
@@ -153,17 +154,21 @@ public interface SqlConditionCriteria extends KeyMapper{
                         if (clzName == null)
                             clzName = alia;
                         Parsed parsed = Parser.get(clzName);
-                        if (BeanUtilX.isBaseType_0(arr[1],x.getValue(),parsed)){
+                        if (BaseTypeUtil.isBaseType_0(arr[1],x.getValue(),parsed)){
                             ite.remove();
                         }else{
-                            BeanUtilX.testNumberValueToDate(parsed.getElement(arr[1]).clz,x);
+                            NumberValueAndDateUtil.testNumberValueToDate(parsed.getElement(arr[1]).clz,x);
+                            if (x.getValue() == null)
+                                ite.remove();
                         }
                     }else{
                         Parsed parsed = criteria.getParsed();
-                        if (BeanUtilX.isBaseType_0(key,x.getValue(),parsed)){
+                        if (BaseTypeUtil.isBaseType_0(key,x.getValue(),parsed)){
                             ite.remove();
                         }else{
-                            BeanUtilX.testNumberValueToDate(parsed.getElement(x.getKey()).clz,x);
+                            NumberValueAndDateUtil.testNumberValueToDate(parsed.getElement(x.getKey()).clz,x);
+                            if (x.getValue() == null)
+                                ite.remove();
                         }
                     }
                 }else if (p == PredicateAndOtherScript.IN
@@ -174,24 +179,15 @@ public interface SqlConditionCriteria extends KeyMapper{
                         continue;
 
                     if (key.contains(".")){
-                        String[] arr = key.split("\\.");
-                        String alia = arr[0];
-                        String clzName = criteria.getAliaMap().get(alia);
-                        if (clzName == null)
-                            clzName = alia;
-                        Parsed parsed = Parser.get(clzName);
-
-                        if (BeanUtilX.isBaseType_0(arr[1],valueList.get(0),parsed)){
+                        if (BaseTypeUtil.isBaseType_0(key,valueList.get(0),criteria)){
                             ite.remove();
                         }
-
                     }else{
                         Parsed parsed = criteria.getParsed();
-                        if (BeanUtilX.isBaseType_0(key,valueList.get(0),parsed)){
+                        if (BaseTypeUtil.isBaseType_0(key,valueList.get(0),parsed)){
                             ite.remove();
                         }
                     }
-
                 }
                 List<X> subList = x.getSubList();
                 if (subList == null || subList.isEmpty())
