+++ x7-spring-boot-starter/src/main/java/io/xream/x7/EnableX7Repository.java
@@ -27,6 +27,10 @@ import java.lang.annotation.*;
 public @interface EnableX7Repository {
 
     /**
+     * Spring JdbcTemplate support baseType bean, while ORM framework not support baseType bean
+     */
+    boolean beanWithoutBaseType() default false;
+    /**
      * t_
      */
     String mappingPrefix() default "";
